Gain Setting using AD5270 driver for INA333
Required changing the gain to a float since the low gain points are
close together.

+++ geoscope/ADCModule.h
@@ -20,7 +20,7 @@ void adcSetup();
 void adcPoll();
 void samplingEnable();
 void samplingDisable();
-void changeAmplifierGain(int);
+void changeAmplifierGain(float);
 void gainSave();
 void gainLoad();
